Move format_study_mode from utils to programmes
since this is more closely tied to the model system in programmes
